theme toggling working
